+++ src/app/(list)/[teamId]/task-lists/[listId]/_components/index.ts
@@ -1,0 +1,4 @@
+export { default as TaskCard } from "./task-card";
+export { default as TaskListNav } from "./task-lists-nav";
+export { default as TaskNav } from "./task-nav";
+export { default as TasksContainer } from "./tasks-container";
