+++ src/app/(list)/[teamId]/task-lists/layout.tsx
@@ -10,7 +10,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <div className="mx-auto mt-24 w-full">
+    <div className="mx-auto mt-24 w-full max-w-1200">
       <header className="mb-27">
         <h1 className="text-20-700 text-text-primary">할 일</h1>
       </header>
